requirement.txt
Django==2.1.5
django-cleanup==3.2.0
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.1
mysqlclient==1.4.1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
six==1.12.0
